Added to stealth functionality
Stealth now (partially) makes individual modules on the ship slightly
transparent as well as the entire ship itself. Currently, some parts of
the modules don't go transparent, and not all of the modules go
transparent.
Current implementation has a directive listening for space bar input
which calls setAlpha for each ship module. Alternate solution is to have
the stealth module get a list of all the modules on the ship and use
dynamic cast to make all the ship modules transparent. This way, only
the stealth module will have a directive function that takes care of all
other modules.