Added timer to Stealth module
The ship will come out of stealth mode automatically after 5 seconds.
Currently, this mechanic can be abused by holding down the space bar,
but this can be fixed by adding in energy consumption. Also fixed the
toggling of stealth for a Turret and Radar. TODO: Energy Consumption &
Tests and Refactoring.