Allow Chunks to Track Score
Added a field to GameObjects that allow them to keep track of score.
Currently, there is no way to increment or change this value. This will
probably be renamed to something else to keep track of kills and/or
captures, etc.
Planning to get this score value from all the game objects in the
universe and total them up to display in the scoreboard.