Added scoreboard
Added a "scoreboard" that shows the amount of score the current player
has. This is currently implemented as un unclickable button. At the
moment, the values it shows are rather meaningless.
Also, added a function to increase the score of a ship, which is
currently a spike solution that increases the score every time the
turret is fired.