Implemented Grapple Weapon
A grapple weapon is currently a laser weapon that, instead of damaging
the body that it hits, grapples to it. This is done by applying a force
to your ship in the direction of the ship that we are grappling to. We
still need to add this weapon to a module via the blueprints, but have
tested it by changing LaserWeapon and it works