Added graphical component of scoreboard
There are now four sections of the scoreboard, one for each team. They
are color coded to correspond to the colors associated with each team on
the main menu. Still need to add logic to update values on the
scoreboard when necessary. Also, is there a better way to format the
text we have on the scoreboard?