Show Money in HUD
Added the ability for a player to see their money in the HUD, in the
same location where they see their score. Also changed the scoreboard to
have buttons be the column labels. Can be used in future to sort columns
maybe? Still need to add logic to update values in scoreboard