Added HUD display to show weapon toggles
Shows which weapons are enabled or disabled using the weapon control
groups. Need to see if we can get names of weapons and check how many
the ship has instead of always showing 4 groups, when the ship may not
have 4.
Also refactored code that hides all HUD displays when the scoreboard is
toggled