C++17
Ballistic weapons aren't working. It's not clear whether the problem is with the blueprint or with code. If you fire a ballistic weapon, and enable debug view, you can briefly see the projectile appear on the ship. It could be a rapid expire time. Consider debugging the expire code of a ballistic projectile.